Misc cleanup
Adjusted Table of Contents using the Markdown All in One plugin.
Renamed SpiderList that seems arbitrary? I see this in the SpiderMan example in Github but nothing in Livebook is initializing that as a module. I know it's pattern matching but it's so weird to match to nothing? Library internals don't use that to my knowledge.
Removing the bunk zero-width space section from the elixir jobs notebook and moving it to the oddities. It's useful to push these into their own notebook so that we can illustrate workarounds.