Job application parsing
Skeleton created for the application field parsing. We want to capture the names and possible values as we do capture this locally as well. Usually I would translate this by hand.

Breezy is already a problem because it looks like the page is in Angular so the HTML and JS is heavily angular-centric. You can parse the questionnaire if you're find inspecting the JSON used to create it through regex. I'm hoping the others are less problematic but the job of the notebook is to also capture stuff like this.