Readme updates
Add new livebook files and sections. Looking to merge this and the primary livebook file in some capacity. It is helpful to duplicate it or have the `Run in Livebook` button displayed. I don't know how useful that button would be on the individual notebooks. It's worth playing around with.