Job description first pass complete
Basic completion though not too happy with the `\[ \]` inserted from the HTML.
This is extremely rough as files should be saved and skipped if present with an override toggle.
The modules should likely exist in their own mix project since there will be overlap between this and the job field parsing.