Job description cleanup
Our version to handle Breezy HR or at the very least **all of the current DockYard jobs** is complete.

This is a bit of a mess with how things are being passed around. We have the url as input, a possible extension, the contents for saving with a possible overwrite option and a read. Reading currently doesn't care about the extension but checking if it exists or saving does matter. Reading could care in the future. It probably makes sense to have a few high level functions control the flow as it exists with far less piping on the outside. The API isn't difficult to understand that you need the path to use it but since it's a tuple and not a dedicated struct, it's far less intuitive. It also makes sense to have escape hatches with things like default function calls with delegating to anonymous functions when you want. That's in case you wanted to switch out the `req` usage for something else.

Basically, it is easy to see this going in multiple directions if I'm not careful. I think it makes sense for the module to encapsulate all the reusability but it's a little clumsy right now and will be until we start to abstract different ATS types. I think we're making the right progress but I can easily see how refactoring is already annoying here over something like VS Code dedicated to an Elixir project.