Refactor read behavior
Removed the need to pipe `Utilities.Persistence.get_path(url_value)` everywhere but I'm not sure I like the alternative.
If the file exists, our read function just pulls it off the disk. If the file doesn't exist, we use the previous `req` usage to get and save the contents. There's probably a more idiomatic way to do this other than an if/else but that also works.
I was looking at passing in a possible function on read to override the behavior if the usage of req was somehow flawed. Defaulting function arguments to an existing function with the same arguments probably wasn't working as I expected. I'm sure there's a way to delegate a possible anonymous function.