Description transforms
Refactor logic from Notebook to complete the BreezyHR transformations. I'm not the biggest fan of the approach as I wasn't clear on things like how the title is part of the bigger HTML with the description only pulling out the part we care about.

`find_description` returns the list with the idea being we would do a detection, return the partial data set we needed then transform that Floki result further but that got thrown to the wind with the inclusion of the title. There is likely ways of compiling patterns I'm overlooking or there's a better way to handle the data flow to inject a title as the very first line.