ElixirForum Jobs | First pass
Gather the `date`, `source_url`, and `page_number` fields.
The `tags` field is roughly the workplace or type like remote, onsite, full-time-contract. Parsing will likely require mapping known tags to fields. Jobs can have more than one tag which is also true of other boards.
The `info` variable has the title, company, location, and sometimes remote jumbled in a mess so we set the title to this for now. It's unclear how I wish to clean this in the future as the post itself should have more unique values to parse correctly.