Add Larajobs Notebook
Since I'm entertaining looking for other Laravel jobs, and it would be useful to show an example for parsing a single squirrelly page of results, I thought it would be useful to have this notebook as another example. There are other Laravel job boards I intend to cover. A few of the Laravel boards seem to pass all of the HTML and "page" through results using Javascript to show or hide as necessary.