Job description: SNAFU
I wanted to capture https://mikezornek.com/posts/2022/8/elixir-consulting-availability/ to have an understanding of how a veteran Elixir developer promotes himself.

This illustrates how fragmented the underlying architecture is but it is also a chance to introduce behaviours learned in a recent Exercism exercise. It would also be ideal to write tests ASAP for the utilities section because testing will help with weird APIs like this.

It's weird to use `find_description` to return an atom, to only then use `transform_description` according to the detection. Then I realize it does make sense as we can't determine the type we need to transform. It should be possible to run one function that detects, delegates to the appropriate module, then returns the atom type, title, and content. The current implementation is definitely clumsy but I do like to refactor this way, once I notice the warts.