ElixirConf 2022 (Virtual)
Add Thursday and Friday templates. This highlights a need to use a notebook as a template. Does branching cover this? Think about repository templates in Github, can that be applied here?

Changing the topic does weirdly remove what is currently multiple blank markdown sections. This is fine but this is why there is TODO placeholder text.