ElixirConf 2022 (Virtual)
Unified understanding section and added VOD notes for the track 1 talks I watched. I really should've recapped but it also highlighted that this is 8 hours * 3 almost. That's not sustainable to recap overnight, it only is if I stick to the talks I missed and I didn't do that way too easily.

I do need a better way to call out what I want to focus on. These next days could be harder given everyone will be home. I really need to focus on the afternoon talks anyway and just the ones I missed first. Let's see if these next few days are better.