ElixirConf 2022 (Virtual) | Friday notes
Notes from Friday talks
I didn't take notes during the last keynote largely due to the technical difficulties but also it was dinner time and tough to take notes. I plan on making another pass through this to have a blog post up by Wednesday which is a little late but I want to catch up on all of these talks I missed.