Larajobs DOM changes
Realized we could gather all anchor tags with the .job-link class
Date is no longer a date string OR an Apply button, we can just capture the date now
DOM changed to literal div hell with absolutely no semantic relation between fields. Thankfully we can use nth-of-type, nth-child, or last-child to slice things down how we need it. I feel like this is a 301 level web scraping course in college.