CodeBEAM US 2022 (Virtual) | Thursday initial skeleton created
Notes from Thursday so far.
CodeBEAM doesn't have a concept of track 1 and 2 but `topics` are `tracks` in the UI so it may be useful to unify all of this.
I should've done this the night before as I missed talks to work on this.
Add Teller challenge skeleton as I suspect this challenge may or may not be different from ElixirConf. The url was given out for any virtual or in person attendees to complete. It'll be worth going over this from scratch again, attempting a new approach and some of what I've learned at the academy.
Captured contacts for sponsors as I have a feeling I should be talking to them. I missed the meet & greet section, hoping it was over multiple days. I believe I can click the "I'm interested" and get something going.