Teller Bank Challenge v1.2 for CodeBEAM America 2022
This challenge deviates from the earlier version slightly. There is a single OTP code (`001137`) sent by the service rather than an internal function, the `f_token` is generated using sha3_512 instead of sha3_256 but I reused quite a bit from the earlier implementation. Now that submissions are closed, we can publish this.