Add an oddity uncovered through DockYard Academy
During a pairing session we discussed this awkward behavior that stumped us at the time. Turns out the warning gives the problem away. While you can technically write code with a module attribute on multiple lines, the compiler emits a warning instead of an error. We can make the tests pass in spite of the warnings too. This deserves a blog post even if it admits a major gaffe we overlooked.