Larajobs | Fix date sorting
Fixed date comparisons as I don't recall if any of the other notebooks really use a DateTime sigil and their sorting just worked(tm).

I'm legit having an "I wrote this?" moment for the `take` anonymous function. I would've thought a Stream would've been cooler but we're talking < 100 items so it really doesn't matter.

It also took me a second to remember we have a useful map/struct thing already so sorting and then turning a map into a list of values means we should sort first, then convert. It was much harder trying to reason about sorting the first value in the list, especially since `Enum.at()` didn't seem to work as I expected. I'd get the dates but it's like the comparison would happen from right to left so 02-07 < 01-31 because 31 is greater than 07. Sorting was just "weird" for lack of a better word and it felt like asc or desc did nothing, which hinted at there being issues with how I was going about my comparisons.