Fix paths for huggingface
It's difficult to reason out how to have a different path locally vs in the docker image.
The `PWD` on huggingface is `/data` which matches the `LIVEBOOK_DATA_PATH` environment variable.

Hopefully setting this in the .env to use with direnv will do what I expect. I haven't killed the Livebook process to test yet.