Elixir Companies app :: Displaying the table
Initial setup to save files to the PWD directory.
Fixed a few enum discard steps as I no longer needed to pull out values that were likely always empty. I never mentioned in the source why I did this.
Having issues with the data transform Smart Cell now and I think it's data related or I'm not setting things up as expected. It could be having issues with spider man as it tries to detect compatible data. For the larajobs app it correctly detects `items` though that's not explicitly defined.