Fix environment
Setting these variables locally was kind of a big problem as now my `autosaves` for May 18 are here instead of the default location(s) where I can generally manage them better. This isn't the worst thing as I could keep autosaves local to the project and use git as a poor man's backup. I may end up doing that and reverting but this will also hinge on the GitHub issue I'm about to raise, trying to determine the DATA_PATH at runtime.